requirements.txt
# Основные зависимости проекта
Flask
python-dotenv
google-generativeai
requests
hubspot-api-client
redis
gunicorn

# --- LlamaIndex и все необходимые плагины ---
pinecone-client
llama-index
llama-index-embeddings-gemini
llama-index-llms-openrouter
llama-index-vector-stores-pinecone
llama-index-postprocessor-sentence-transformers